create features for every scheme, upgrade documentation
