+++ src/util.rs
@@ -1,6 +1,8 @@
+use crate::Compressable;
 use group::{ff::Field, Group};
 use irmaseal_curve::{G1Projective, G2Projective, Gt, Scalar};
 use rand::{CryptoRng, RngCore};
+use subtle::CtOption;
 use tiny_keccak::Hasher;
 
 /// Size of a compressed target group element.
@@ -14,6 +16,9 @@ pub(crate) const G2_BYTES: usize = 96;
 
 /// Size of a serialized scalar.
 pub(crate) const SCALAR_BYTES: usize = 32;
+
+/// Size of the (default) identity buffer.
+pub(crate) const ID_BYTES: usize = 64;
 
 #[inline(always)]
 pub fn rand_scalar<R: RngCore + CryptoRng>(rng: &mut R) -> Scalar {
@@ -71,6 +76,46 @@ pub fn shake256<const N: usize>(slice: &[u8]) -> [u8; N] {
     digest.finalize(&mut buf);
 
     buf
+}
+
+/// Byte representation of an identity.
+/// Most schemes (not all) use the same representation.
+///
+/// This identity is obtained by hashing using sha3_512.
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub struct Identity(pub [u8; ID_BYTES]);
+
+impl Identity {
+    /// Hash a byte slice to a set of Identity parameters, which acts as a user public key.
+    /// Uses sha3-512 internally.
+    pub fn derive(b: &[u8]) -> Identity {
+        Identity(sha3_512(b))
+    }
+
+    /// Hash a string slice to a set of Identity parameters.
+    /// Directly converts characters to UTF-8 byte representation.
+    pub fn derive_str(s: &str) -> Identity {
+        Self::derive(s.as_bytes())
+    }
+
+    /// Create a scalar from an identity.
+    #[allow(unused)]
+    pub fn to_scalar(&self) -> Scalar {
+        Scalar::from_bytes_wide(&self.0)
+    }
+}
+
+impl Compressable for Gt {
+    const OUTPUT_SIZE: usize = GT_BYTES;
+    type Output = [u8; Self::OUTPUT_SIZE];
+
+    fn to_bytes(&self) -> [u8; GT_BYTES] {
+        self.to_compressed()
+    }
+
+    fn from_bytes(bytes: &[u8; GT_BYTES]) -> CtOption<Self> {
+        Self::from_compressed(bytes)
+    }
 }
 
 #[cfg(test)]
